Rename equivalent_width.py to flux.py
