+++ specutils/analysis/__init__.py
@@ -1,6 +1,6 @@
 from __future__ import absolute_import
 
-from .equivalent_width import line_flux, equivalent_width  # noqa
+from .flux import line_flux, equivalent_width  # noqa
 from .snr import snr  # noqa
 from .centroid import centroid  # noqa
 from .width import gaussian_sigma_width, gaussian_fwhm, fwhm # noqa
