+++ specutils/analysis/flux.py
